insercion en postgres

+++ postgres/creacion_tablas.sql
@@ -18,7 +18,7 @@ create table centros_de_datos (
 
 create table usuarios (
     id          serial,
-    id_centro   numeric(2),
+    id_centro   integer,
     nombre      varchar(30),
     constraint pk_usuarios primary key (id),
     constraint fk_usuarios_centros foreign key (id_centro) references centros_de_datos (id)
@@ -27,7 +27,7 @@ create table usuarios (
 create table logueos (
     id          serial,
     fechahora   timestamp,
-    id_usuario  numeric(3),
+    id_usuario  integer,
     ticket      varchar(400),
     csfr        varchar(100),
     constraint pk_logueos primary key (id, fechahora),
@@ -36,7 +36,7 @@ create table logueos (
 
 create table nodos (
     id          serial,
-    id_centro   numeric(2),
+    id_centro   integer,
     nombre      varchar(30),
     constraint pk_nodos primary key (id),
     constraint fk_nodos_centros foreign key (id_centro) references centros_de_datos (id)
@@ -44,7 +44,7 @@ create table nodos (
 
 create table maquinas_virtuales (
     id          serial,
-    id_nodo     numeric(2),
+    id_nodo     integer,
     nombre      varchar(30),
     constraint pk_maquinas_virtuales primary key (id),
     constraint fk_maquinas_nodos foreign key (id_nodo) references nodos (id)
@@ -52,7 +52,7 @@ create table maquinas_virtuales (
 
 create table contenedores (
     id          serial,
-    id_nodo     numeric(2),
+    id_nodo     integer,
     nombre      varchar(30),
     constraint pk_contenedores primary key (id),
     constraint fk_contenedores_nodos foreign key (id_nodo) references nodos (id)
@@ -60,7 +60,7 @@ create table contenedores (
 
 create table plantillas (
     id          serial,
-    id_nodo     numeric(2),
+    id_nodo     integer,
     nombre      varchar(30),
     constraint pk_plantillas primary key (id),
     constraint fk_plantillas_nodos foreign key (id_nodo) references nodos (id)
@@ -76,9 +76,9 @@ create table tipos_operaciones (
 create table operaciones (
     id          serial,
     fechahora   timestamp,
-    id_usuario  numeric(3),
-    id_nodo     numeric(2),
-    id_tipo     numeric(3),
+    id_usuario  integer,
+    id_nodo     integer,
+    id_tipo     integer,
     constraint pk_operaciones primary key (id,fechahora),
     constraint fk_operaciones_usuario foreign key (id_usuario) references usuarios (id),
     constraint fk_operaciones_nodo foreign key (id_nodo) references nodos (id),
